+++ plugins/mongodb/mongodb_data.go
@@ -89,7 +89,7 @@ func (d *MongodbData) addStat(acc plugins.Accumulator, statLine reflect.Value, s
 }
 
 func (d *MongodbData) add(acc plugins.Accumulator, key string, val interface{}) {
-	acc.AddFieldsWithTime(
+	acc.AddFields(
 		key,
 		map[string]interface{}{
 			"value": val,